se agrega creacion y recupero de patogenos

+++ src/test/java/ar/edu/unq/eperdemic/utils/MockitoTest.kt
@@ -15,6 +15,6 @@ class MockitoTest {
         Mockito.`when`(especieMock.id).thenReturn(42)
         val patogenoMock = Mockito.mock(Patogeno::class.java)
         Mockito.`when`(patogenoMock.crearEspecie("sarasa", "sarasa")).thenReturn(especieMock)
-       Assert.assertEquals(42.toLong(), patogenoMock.crearEspecie("sarasa", "sarasa").id)
+       Assert.assertEquals(42, patogenoMock.crearEspecie("sarasa", "sarasa").id)
     }
 }
